+++ lib/classes/class-utility.php
@@ -72,18 +72,33 @@ namespace wpCloud\StatelessMedia {
        * @return string
        */
       public static function randomize_filename( $filename ) {
+        $return = apply_filters('stateless_skip_cache_busting', null, $filename);
+        if($return){
+          return $return;
+        }
         $info = pathinfo($filename);
         $ext = empty($info['extension']) ? '' : '' . $info['extension'];
         $_parts = array();
+        $rand = substr(md5(time()), 0, 8);
+
+        $body_rewrite_types = ud_get_stateless_media()->get( 'sm.body_rewrite_types' );
+        if(empty($info['extension']) || strpos($body_rewrite_types, $info['extension']) === false){
+          return $filename;
+        }
+
+        if(strpos($filename, $rand) !== false){
+          return $filename;
+        }
+
         if (strpos($info['filename'], '@')) {
           $_cleanName = explode('@', $info['filename'])[0];
           $_retna = explode('@', $info['filename'])[1];
-          $_parts[] = substr(md5(time()), 0, 8);
+          $_parts[] = $rand;
           $_parts[] = '-';
           $_parts[] = strtolower($_cleanName);
           $_parts[] = '@' . strtolower($_retna);
         } else {
-          $_parts[] = substr(md5(time()), 0, 8);
+          $_parts[] = $rand;
           $_parts[] = '-';
           $_parts[] = strtolower($info['filename']);
         }
@@ -150,8 +165,13 @@ namespace wpCloud\StatelessMedia {
        * @param $force Whether to force the upload incase of it's already exists.
        * @return bool|string
        */
-      public static function add_media( $metadata, $attachment_id, $force = false ) {
+      public static function add_media( $metadata, $attachment_id, $force = false, $args = array() ) {
+        global $stateless_synced_full_size;
+        $file = '';
         $upload_dir = wp_upload_dir();
+        $args = wp_parse_args($args, array(
+          'no_thumb' => false,
+        ));
 
         /* Get metadata in case if method is called directly. */
         if( current_filter() !== 'wp_generate_attachment_metadata' && current_filter() !== 'wp_update_attachment_metadata' ) {
@@ -165,10 +185,17 @@ namespace wpCloud\StatelessMedia {
           $fullsizepath = wp_normalize_path( get_attached_file( $attachment_id ) );
           // Make non-images uploadable.
           if( empty( $metadata['file'] ) && $attachment_id ) {
-            $metadata = array( "file" => str_replace( trailingslashit( $upload_dir[ 'basedir' ] ), '', get_attached_file( $attachment_id ) ) );
+            $file = str_replace( wp_normalize_path(trailingslashit( $upload_dir[ 'basedir' ] )), '', $fullsizepath );
+            if(empty($metadata)){
+              $metadata = array();
+            }
+            $mime_type = get_post_mime_type( $attachment_id );
+            if($mime_type != "application/pdf"){
+              $metadata["file"] = $file;
+            }
           }
 
-          $file = wp_normalize_path( $metadata[ 'file' ] );
+          $file = wp_normalize_path( !empty($metadata[ 'file' ])?$metadata[ 'file' ]:$file );
 
           $image_host = ud_get_stateless_media()->get_gs_host();
           $bucketLink = apply_filters('wp_stateless_bucket_link', $image_host);
@@ -178,7 +205,7 @@ namespace wpCloud\StatelessMedia {
             "height" => isset( $metadata[ 'height' ] )  ? $metadata[ 'height' ] : null,
             'object-id' => $attachment_id,
             'source-id' => md5( $attachment_id.ud_get_stateless_media()->get( 'sm.bucket' ) ),
-            'file-hash' => md5( $metadata[ 'file' ] )
+            'file-hash' => md5( $file )
           );
 
           if($attachment_id && !empty($metadata) && !$force){
@@ -190,7 +217,7 @@ namespace wpCloud\StatelessMedia {
 
           /* Add default image */
           $media = $client->add_media( $_mediaOptions = array_filter( array(
-            'force' => $force,
+            'force' => $force && $stateless_synced_full_size != $attachment_id,
             'name' => $file,
             'absolutePath' => wp_normalize_path( get_attached_file( $attachment_id ) ),
             'cacheControl' => $_cacheControl = self::getCacheControl( $attachment_id, $metadata, null ),
@@ -235,10 +262,10 @@ namespace wpCloud\StatelessMedia {
           }
 
           /* Now we go through all available image sizes and upload them to Google Storage */
-          if( !empty( $metadata[ 'sizes' ] ) && is_array( $metadata[ 'sizes' ] ) ) {
+          if( !empty( $metadata[ 'sizes' ] ) && is_array( $metadata[ 'sizes' ] ) && $args['no_thumb'] != true ) {
 
             $path = wp_normalize_path( dirname( get_attached_file( $attachment_id ) ) );
-            $mediaPath = wp_normalize_path( trim( dirname( $metadata[ 'file' ] ), '\/\\' ) );
+            $mediaPath = wp_normalize_path( trim( dirname( $file ), '\/\\' ) );
 
             foreach( (array) $metadata[ 'sizes' ] as $image_size => $data ) {
 
@@ -286,12 +313,15 @@ namespace wpCloud\StatelessMedia {
           }
 
           // Stateless mode: we don't need the local version.
-          if(ud_get_stateless_media()->get( 'sm.mode' ) === 'stateless'){
+          if(ud_get_stateless_media()->get( 'sm.mode' ) === 'stateless' && $args['no_thumb'] != true){
             unlink($fullsizepath);
           }
 
           update_post_meta( $attachment_id, 'sm_cloud', $cloud_meta );
 
+          if($args['no_thumb'] == true){
+            $stateless_synced_full_size = $attachment_id;
+          }
         }
 
         return $metadata;
